Move assignRedeemers to separate module
