+++ lib/wallet/cardano-wallet.cabal
@@ -410,6 +410,7 @@ library
     Cardano.Wallet.Write.Tx
     Cardano.Wallet.Write.Tx.Balance
     Cardano.Wallet.Write.Tx.Gen
+    Cardano.Wallet.Write.Tx.Redeemers
     Cardano.Wallet.Write.Tx.TimeTranslation
     Cardano.Wallet.Write.UTxOAssumptions
     Control.Concurrent.Concierge
